+++ mizani/tests/test_palettes.py
@@ -8,7 +8,8 @@ import numpy.testing as npt
 from ..palettes import (hls_palette, husl_palette, rescale_pal,
                         area_pal, abs_area, grey_pal, hue_pal,
                         brewer_pal, gradient_n_pal, cmap_pal,
-                        desaturate_pal, manual_pal, xkcd_palette)
+                        desaturate_pal, manual_pal, xkcd_palette,
+                        crayon_palette)
 
 
 def test_hls_palette():
@@ -149,5 +150,11 @@ def test_manual_pal():
 
 def test_xkcd_palette():
     values = xkcd_palette(['apple green', 'red', 'tan brown'])
+    assert len(values) == 3
+    assert all(s[0] == '#' and len(s) == 7 for s in values)
+
+
+def test_crayon_palette():
+    values = crayon_palette(['banana mania', 'red', 'yellow'])
     assert len(values) == 3
     assert all(s[0] == '#' and len(s) == 7 for s in values)